requirements.txt
# Core Dependencies
dspy-ai>=2.4.0
openai>=1.0.0
pydantic>=2.0.0
litellm>=1.50.0

# Machine Learning & NLP
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
scipy>=1.10.0
numpy>=1.24.0

# Additional dependencies for DSPy
transformers>=4.20.0
torch>=1.10.0

# PDF Processing Dependencies (preprocessing module)
google-generativeai>=0.3.0    # Gemini Vision API
llmwhisperer-client>=0.14.0    # LLMWhisperer for high-quality text extraction
mineru>=0.1.0                  # Layout detection and figure extraction
Pillow>=9.0.0                  # Image processing
loguru>=0.7.0                  # Enhanced logging

# Optional: For enhanced functionality
requests>=2.28.0
pathlib-mate>=1.0.0

# Development Dependencies (optional)
pytest>=7.0.0
black>=23.0.0
isort>=5.12.0
mypy>=1.0.0

# Environment Management
python-dotenv>=1.0.0